Moves are now mostly culled correctly when in check and also correctly prevent the king from moving into a check.

+++ src/game.h
@@ -24,7 +24,7 @@ typedef enum Direction {
 int16_t get_piece(uint32_t *game_state, uint16_t loc);
 uint32_t check_game_condition(uint16_t *game_condition, uint16_t piece, int32_t location);
 void set_piece_moved(uint16_t *game_condition, uint16_t piece, int32_t location);
-int32_t piece_in_direction(uint32_t *game_state, int32_t loc, Direction direction);
+int16_t piece_in_direction(uint32_t *game_state, int32_t loc, Direction direction);
 uint16_t check_valid_move(uint32_t *game_state, uint16_t *game_condition, uint64_t *legal_moves, int32_t loc, int32_t dest);
 void update_legal_moves(uint32_t *game_state, uint16_t *game_condition, uint64_t *legal_moves);
 void move_piece(uint32_t *game_state, uint16_t loc, uint16_t dest);
